breaking out functions: finished occasions component and login
