--- admin-site/src/ButtonActions.svelte
@@ -1,9 +1,0 @@
-<script context="module">
-
-export function goBackAPage(pageState){
-    pageState --;
-    // not sure this is the right space for this
-    // broadcastStatus = "unsent"
-  }
-
-</script>
